requirements.txt
# Data collection & mining
PyDriller>=2.6
PyGithub>=2.1.1
requests>=2.31.0

# NLP & embeddings
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2

# ML & modeling
xgboost>=2.0.0
scikit-learn>=1.3.0
scikit-optimize>=0.9.0  # Bayesian optimization
shap>=0.43.0

# Data processing
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
joblib>=1.3.0
pyyaml>=6.0.1
